Basic updates to enemy

+++ Prodigium/GUIHandler.cpp
@@ -351,8 +351,8 @@ void GUIHandler::RenderTrapGUI()
 void GUIHandler::RenderBrainGUI(int playerHp, int clues)
 {
     float fade = 1.0f;
-    float hp = playerHp;
-    fade = std::max(std::min(hp, 100.0f), 10.0f) * 0.01f;
+    int hp = playerHp;
+    fade = std::max(std::min(hp, 100), 10) * 0.01f;
 
     bool* isActive = new bool(true);
     SetNextWindowPos(ImVec2((float)Graphics::GetWindowWidth() - 250, 25));
